Create Reada2D-Array.java
Read a 2D-array
Write a function to read values of a number of rows, number of columns, 2 dimensional (2D) array elements and display the 2D array in a matrix form.

 

Input 
    2 
    3 
    1 4 5 
    2 3 0 

  

    Where, 

First line of represents the number of rows.  
Second line of input represents the number of columns.  
Third line contains array elements of the 1st row and so on. 
 

Output 
    1 4 5 
    2 3 0

 

    where

There must be single space between 2 numbers in the row.
There should not be any space after the last number in the row. e.g. in the 2nd row, there should not be any space after number 0. 
 

Assume that, 

Row and column values are integers within the range [1 to 100]. 
2D Array elements are within the range [-2147483648 to 2147483647].