Create greatestNumberFunctionTemplate.java
Given N1 and N2 of type template, Find the greatest number using function template.

 

Write a function 'solution' that accepts N1 and N2 as parameters and returns the greatest number of template type.
 

Input
    2
    2.5
    6.11 
    
    Where,

The first line contains the data type which can be 1, 2 or 3. i.e. 1 for an integer, 2 for a float, 3 for a character.
The second line contains input N1.
The third line contains input N2.
    
Output
    6.11